Added delete prompts, closes issue #35

+++ app/helpers/application_helper.rb
@@ -23,18 +23,18 @@ module ApplicationHelper
       link_to('<em class="fa fa-exclamation-triangle"></em>'.html_safe, recover_algorithm_path(algorithm), method: :post, class: 'btn btn-primary')
     when :published
       html = link_to('<em class="fa fa-pencil"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-primary')
-      html << link_to('<em class="fa fa-trash"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-danger', method: :delete)
+      html << link_to('<em class="fa fa-trash"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-danger', method: :delete, data: { confirm: 'Are you sure?' })
       return html
     when :unpublished_changes
       html = link_to('<em class="fa fa-pencil"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-primary')
       html << link_to('<em class="fa fa-undo"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-warning') #TODO implement undo path
-      html << link_to('<em class="fa fa-trash"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-danger', method: :delete)
+      html << link_to('<em class="fa fa-trash"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-danger', method: :delete, data: { confirm: 'Are you sure?' })
       return html
     when :validating, :creating, :testing
       '<i class="fa fa-refresh fa-spin fa-fw" aria-hidden="true"></i>'.html_safe
     when *Algorithm.wizard_steps
       html = link_to('<em class="fa fa-chevron-circle-right"></em>'.html_safe, algorithm_algorithm_wizard_path(algorithm, algorithm.status), class: 'btn btn-primary')
-      html << link_to('<em class="fa fa-trash"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-danger', method: :delete)
+      html << link_to('<em class="fa fa-trash"></em>'.html_safe, algorithm_path(algorithm), class: 'btn btn-danger', method: :delete, data: { confirm: 'Are you sure?' })
       return html
     end
   end
